requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
pygame==2.1.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
